requirements.txt
flask
fuzzywuzzy==0.18.0
google-generativeai
python-dotenv==0.19.1
werkzeug==2.2.2
python-Levenshtein
gunicorn==19.7.1
